Moved some files to util/
