+++ AddressBook.cpp
@@ -9,7 +9,7 @@
 #include <boost/lexical_cast.hpp>
 #include <cryptopp/osrng.h>
 #include "base64.h"
-#include "util.h"
+#include "util/util.h"
 #include "Identity.h"
 #include "Log.h"
 #include "NetDb.h"
