+++ AddressBook.h
@@ -9,7 +9,7 @@
 #include <mutex>
 #include <boost/asio.hpp>
 #include "base64.h"
-#include "util.h"
+#include "util/util.h"
 #include "Identity.h"
 #include "Log.h"
