+++ DaemonLinux.cpp
@@ -9,7 +9,7 @@
 #include <sys/stat.h>
 
 #include "Log.h"
-#include "util.h"
+#include "util/util.h"
 
 
 void handle_signal(int sig)
