+++ DaemonWin32.cpp
@@ -1,5 +1,5 @@
 #include "Daemon.h"
-#include "util.h"
+#include "util/util.h"
 #include "Log.h"
 
 #ifdef _WIN32
