+++ Garlic.cpp
@@ -6,7 +6,7 @@
 #include "I2NPProtocol.h"
 #include "Tunnel.h"
 #include "TunnelPool.h"
-#include "Timestamp.h"
+#include "util/Timestamp.h"
 #include "Destination.h"
 #include "Garlic.h"
