+++ HTTPProxy.cpp
@@ -5,7 +5,7 @@
 #include <string>
 #include <atomic>
 #include "HTTPProxy.h"
-#include "util.h"
+#include "util/util.h"
 #include "Identity.h"
 #include "Streaming.h"
 #include "Destination.h"
