+++ Makefile
@@ -29,6 +29,7 @@ tests: mk_build_test_dir $(TESTS)
 mk_build_dir:
 	mkdir -p obj
 	mkdir -p obj/transport
+	mkdir -p obj/util
 mk_build_test_dir:
 	mkdir -p obj/tests
 
@@ -45,6 +46,7 @@ deps:
 	@mkdir -p obj
 	@mkdir -p obj/transport
 	@mkdir -p obj/tests
+	@mkdir -p obj/util
 	$(CXX) $(CXXFLAGS) $(NEEDED_CXXFLAGS) -MM *.cpp > $(DEPS)
 	@sed -i -e '/\.o:/ s/^/obj\//' $(DEPS)
 
@@ -52,6 +54,7 @@ obj/%.o : %.cpp
 	@mkdir -p obj
 	@mkdir -p obj/transport
 	@mkdir -p obj/tests
+	@mkdir -p obj/util
 	$(CXX) $(CXXFLAGS) $(NEEDED_CXXFLAGS) $(INCFLAGS) $(CPU_FLAGS) -c -o $@ $<
 
 # '-' is 'ignore if missing' on first run
