+++ UPnP.cpp
@@ -18,7 +18,7 @@
 #include "RouterContext.h"
 #include "UPnP.h"
 #include "NetDb.h"
-#include "util.h"
+#include "util/util.h"
 
 #include <miniupnpc/miniupnpc.h>
 #include <miniupnpc/upnpcommands.h>
