+++ UPnP.h
@@ -12,7 +12,7 @@
 
 #include <boost/asio.hpp>
 
-#include "util.h"
+#include "util/util.h"
 
 #define I2P_UPNP_TCP 1
 #define I2P_UPNP_UDP 2
